% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyse_FH.R
\name{Analyse_FH}
\alias{Analyse_FH}
\title{Analyse FoundHaplo FH score values for predictions}
\usage{
Analyse_FH(results_FILE, save_FH_output_DIR, critical_percentile = 0.99)
}
\arguments{
\item{results_FILE}{Path to a single .txt file with all the FH scores by FoundHaplo}

\item{save_FH_output_DIR}{Directory to save the graphical output of the FH scores}

\item{critical_percentile}{Critical percentile of the control cohort to derive predictions}
}
\value{
A data frame predictions that gave FH scores greater than a selected critical percentile
}
\description{
Predictions that gave FH scores greater than a selected critical percentile of the 1000 Genomes control cohort will be lislted.
ECDF of the FH scores are plotted in save_FH_output_DIR for further analysis
}
\examples{
orig_DIR <- getwd()
temp_DIR <- tempdir()
setwd( temp_DIR )
write.table(FH_IBD_scores,"FH_IBD_scores.txt",sep = "\t",quote=FALSE, row.names=FALSE,col.names = TRUE)
Analyse_FH("FH_IBD_scores.txt",temp_DIR,0.99)
setwd(orig_DIR)
}
