% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Annotate_ALT.R
\name{Annotate_ALT}
\alias{Annotate_ALT}
\title{Annotate the ALT column of bi-allelic SNPs in a GATK VCF file}
\usage{
Annotate_ALT(input_vcf, dir_input_ref, dir_output, output_file_name)
}
\arguments{
\item{input_vcf}{Path to a VCF file  (type \code{"character"})}

\item{dir_input_ref}{Directory path to ANNOVAR reference .txt files with general MAF  (type \code{"character"})}

\item{dir_output}{Path to a directory to save the output VCF files (type \code{"character"})}

\item{output_file_name}{Name of the annotated vcf file (type \code{"character"})}
}
\value{
Write a valid new VCF file with ALT column annotated
}
\description{
Annotate the ALT column of bi-allelic SNPs in a GATK VCF file
}
\details{
Make sure the input_vcf chromosome column doesn't have the "chr" prefix.
Takes a GATK generated VCF file with homozygous calls and annotate the ALT column using ANNOVAR reference files.
Download ANNOVAR reference .txt files and specify the path to the main folder in "dir_input_ref"
}
\examples{
orig_dir <- getwd()
temp_dir <- tempdir()
setwd( temp_dir )
write.vcf(FAME1_test_cohort_GATK,paste0(temp_dir,"/","FAME1_test_cohort_GATK.vcf.gz"))
write.table(trimmed_chr8_hg19_ALL.sites.2015_08,paste0(temp_dir,"/","hg19_ALL.sites.2015_08.txt"),sep = "\t",quote=FALSE, row.names=FALSE,col.names = FALSE)
Annotate_ALT(input_vcf = paste0(temp_dir,"/","FAME1_test_cohort_GATK.vcf.gz"), # read input from tempdir()
                         dir_input_ref = temp_dir,
                         dir_output = temp_dir,
                         output_file_name = "example_Annotate_ALT")
setwd( orig_dir )
}
