% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_hap_VCF.R
\name{Create_hap_VCF}
\alias{Create_hap_VCF}
\title{Takes a phased input data file with columns in VCF format and separate haplotypes into two columns}
\usage{
Create_hap_VCF(haplotype_file)
}
\arguments{
\item{haplotype_file}{A dataframe file containing imputed SNP genotyped data in VCF format (including columns in a typical VCF file) after removing meta data}
}
\value{
A data frame with haplotypes of test individuals in two columns
}
\description{
sample.A with alleles 0|1 will be split to two columns with sample.A:a=0 and sample.A:b=1
}
\examples{
haplotype_info=Create_hap_VCF(haplotype_file=data.frame(cbind(FAME1_test_cohort@fix,FAME1_test_cohort@gt)))
}
