% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_jobs.R
\name{Create_jobs}
\alias{Create_jobs}
\title{Create manifest.txt to run batch jobs}
\usage{
Create_jobs(
  path_manifest,
  path_test_sample_chunks,
  path_control_sample_chunks,
  DCV,
  minor_allele_cutoff,
  imputation_quality_score_cutoff_test,
  frequency_type,
  dir_geneticMap,
  dir_disease_files,
  test_file,
  test_name,
  dir_controls_file,
  dir_to_save_report,
  dir_TEMP
)
}
\arguments{
\item{path_manifest}{Path of the manifest.txt file}

\item{path_test_sample_chunks}{Path to the .txt files with chunks of test sample IDs}

\item{path_control_sample_chunks}{Path to the .txt files with chunks of control sample IDs}

\item{DCV}{Name of the disease causing variant of interest i.e FAME1.chr8.119379052 (type \code{"character"})}

\item{minor_allele_cutoff}{The minimum minor allele frequncy of SNPs allowed, we recommend this to be 0 (type \code{"numeric"})}

\item{imputation_quality_score_cutoff_test}{Minimum allowed imputation quality which is R-squared. Recommend to use 0.3 if the cohort has >100 samples ; 0 otherwise (type \code{"numeric"})}

\item{frequency_type}{Population of the test cohort i.e one of EUR,AMR,SAS,EAS,AFR etc (type \code{"character"})}

\item{dir_geneticMap}{Directory path to genetic_map_HapMapII_GRCh37 location (type \code{"character"})}

\item{dir_disease_files}{directory of the disease haplotype VCFs for a single disease variant(type \code{"character"})}

\item{test_file}{path of the test cohort file (type \code{"character"})}

\item{test_name}{meaningful name for the test cohort  (type \code{"character"})}

\item{dir_controls_file}{Directory where the 1000genome control files are stored  (type \code{"character"})}

\item{dir_to_save_report}{Directory path to save the required details of the IBD sharing to analyze later  (type \code{"character"})}

\item{dir_TEMP}{Directory path to save the temporary files  (type \code{"character"})}

\item{test_list}{path to a set of 100 test samples from the test cohort  (type \code{"character"})}

\item{data_type}{"test" or "control (type \code{"character"})}
}
\value{
Write a manifest.txt file that includes all the parameters required to run FoundHaplo. Each line of the manifest.txt file can be submitted as a seperate job parallely.
}
\description{
Takes all the input parameters required to run FoundHaplo and create manifest.txt file to parallelise the process
}
\examples{
add example
}
