% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-FAME1_control_cohort.R
\docType{data}
\name{FAME1_control_cohort}
\alias{FAME1_control_cohort}
\title{Example control cohort for FAME1.chr8.119379052.}
\format{
## `FAME1_control_cohort`
}
\usage{
data(FAME1_control_cohort)
}
\description{
Example control cohort simulated from 1000 Genomes phase 3
}
\keyword{datasets}
