% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-FAME1_disease_cohort.R
\docType{data}
\name{FAME1_disease_cohort}
\alias{FAME1_disease_cohort}
\title{Example disease VCF files for FAME1.chr8.119379052.}
\format{
## `FAME1_disease_cohort`
}
\usage{
data(FAME1_disease_cohort)
}
\description{
Example disease VCF files simulated from 1000 Genomes phase 3
}
\keyword{datasets}
