% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-FAME1_test_cohort.R
\docType{data}
\name{FAME1_test_cohort}
\alias{FAME1_test_cohort}
\title{Example test cohort for FAME1.chr8.119379052.}
\format{
## `FAME1_test_cohort`
}
\usage{
data(FAME1_test_cohort)
}
\description{
Example test cohort simulated from 1000 Genomes phase 3
}
\keyword{datasets}
