% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-FH_IBD_scores.R
\docType{data}
\name{FH_IBD_scores}
\alias{FH_IBD_scores}
\title{Example FoundHaplo results for FAME1.chr8.119379052.}
\format{
## `FH_IBD_scores`
}
\usage{
data(FH_IBD_scores)
}
\description{
Example FoundHaplo results for FAME1.chr8.119379052. to be used by Analyse_FH.R
}
\keyword{datasets}
