% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-file00.R
\docType{data}
\name{file00}
\alias{file00}
\title{Example .txt file with 100 samples from FAME1_test_cohort to run FoundHaplo for disease FAME1.chr8.119379052.}
\format{
## `file00`
}
\usage{
data(file00)
}
\description{
Example .txt file with 100 samples from FAME1_test_cohort simulated from 1000 Genomes phase 3
}
\keyword{datasets}
