% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-genetic_map_chr8.R
\docType{data}
\name{genetic_map_chr8}
\alias{genetic_map_chr8}
\alias{genetic_map_GRCh37_chr8}
\title{hg19 genetic maps from HapMap data}
\format{
## `genetic_map_GRCh37_chr8`
}
\usage{
data(genetic_map_GRCh37_chr8)
}
\description{
Example hg19 HapMap genetic map for chromosome 8
}
\keyword{datasets}
