% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert_bp_cM.R
\name{Convert_bp_cM}
\alias{Convert_bp_cM}
\title{Adds a column with centiMorgan values to the input file}
\usage{
Convert_bp_cM(haplotype_file, DCV, geneticMap_DIR)
}
\arguments{
\item{haplotype_file}{A dataframe file including genotype data, with first nine columns in VCF format after removing meta data, if any}

\item{DCV}{Name of the disease causing variant of interest i.e FAME1.chr8.119379052}

\item{geneticMap_DIR}{Directory to genetic_map_HapMapII_GRCh37 location}
}
\value{
A dataframe with cM distance annotated
}
\description{
"position_cM" column will be added after column "POS" in to a data frame file, where first nine columns are in VCF format.
Genome built should be GRCh37 and genetic map in "geneticMap_DIR" must have chromosomes named with "chr" prefix, add the prefix if needed.
}
\examples{
orig_DIR <- getwd()
temp_DIR <- tempdir()
setwd( temp_DIR )
write.table(genetic_map_GRCh37_chr8,"genetic_map_GRCh37_chr8.txt",sep = "\t",quote=FALSE, row.names=FALSE,col.names = TRUE)
haplotype_file_cM=Convert_bp_cM(haplotype_file=data.frame(cbind(FAME1_test_cohort@fix,FAME1_test_cohort@gt)),DCV="FAME1.chr8.119379052",geneticMap_DIR=temp_DIR)
setwd(orig_DIR)
}
