% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find_bp_to_trim.R
\name{Find_bp_to_trim}
\alias{Find_bp_to_trim}
\title{Outputs the base pair positions corresponding to +-10cM from a given locus}
\usage{
Find_bp_to_trim(
  input_vector,
  geneticMap_DIR,
  output_file,
  size_to_trim_cM = 10
)
}
\arguments{
\item{input_vector}{A vector with the variants (type \code{"character"})}

\item{geneticMap_DIR}{Directory to genetic_map_HapMapII_GRCh37 location (type \code{"character"})}

\item{output_file}{File path to save the results (type \code{"character"})}

\item{size_to_trim_cM}{Size of the region allowed around the disease variant, default is 10cM either side (type \code{"numeric"})}
}
\value{
Return a data frame with three columns 1.Original variant 2. base pair value of 10cM to the left of the variant 3. base pair value of 10cM to the right of the variant. The resulting dataframe is also written to the output_file location.
}
\description{
Takes a vector with disease causing variants written in "FAME1.chr8.119379052" format, and calculate base pair values corresponding to +-10cM from the disease variant locus.
Genome built should be GRCh37 and genetic map in "geneticMap_DIR" must have chromosomes named with "chr" prefix, add the prefix if needed.
}
\examples{
orig_DIR <- getwd()
temp_DIR <- tempdir()
setwd(temp_DIR)
write.table(genetic_map_GRCh37_chr8,"genetic_map_GRCh37_chr8.txt",sep = "\t",quote=FALSE, row.names=FALSE,col.names = TRUE)
Find_bp_to_trim(input_vector=c("FAME1.chr8.119379052"),geneticMap_DIR=temp_DIR,output_file=paste0(temp_DIR,"/DCV_bp.txt"))
setwd( orig_DIR )
}
