% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find_bp_to_trim.R
\name{Find_bp_to_trim}
\alias{Find_bp_to_trim}
\title{Outputs the base pair positions corresponding to +-20cM from a given locus}
\usage{
Find_bp_to_trim(input_vector, dir_geneticMap, output_file)
}
\arguments{
\item{input_vector}{A vector with the variants (type \code{"character"})}

\item{dir_geneticMap}{Directory to genetic_map_HapMapII_GRCh37 location (type \code{"character"})}

\item{output_file}{Path to save the resutls (type \code{"character"})}
}
\value{
Return a data frame with three columns 1.Original variant 2. base pair value of 20cM to the left of the variant 3. base pair value of 20cM to the right of the variant
}
\description{
Takes a text file with disease causing variants written in "FAME1.chr8.119379052." format, variants added in new line or a vector of variants in the same format and calculate base pair values corresponding to +-20cM from the disease variant locus
Genome built should be GRCh37 and genetic map in "dir_geneticMap" must have chromosomes named with "chr" prefix, add the prefix if needed
}
\examples{
orig_dir <- getwd()
temp_dir <- tempdir()
setwd( temp_dir )
write.table(genetic_map_GRCh37_chr8,"genetic_map_GRCh37_chr8.txt",sep = "\t",quote=FALSE, row.names=FALSE,col.names = TRUE)
Find_bp_to_trim(input_vector=c("FAME1.chr8.119379052."),dir_geneticMap=temp_dir,output_file=paste0(temp_dir,"/DCV_bp.txt"))
setwd( orig_dir )
}
