% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phasing_by_pedigree.R
\name{Phasing_by_pedigree}
\alias{Phasing_by_pedigree}
\title{Phase samples with a disease variant by pedigrees}
\usage{
Phasing_by_pedigree(input_vcf, output_DIR, sample_info_file, n.cores = 1)
}
\arguments{
\item{input_vcf}{File path to a VCF file  (type \code{"character"})}

\item{output_DIR}{Directory to save the output VCF files. Must have a dedicated directory for the output (type \code{"character"})}

\item{sample_info_file}{File path to a tab delimited .txt file with sample names and type of phasing to be used. Each phasing is included in a new line, include sample names as in the VCF file in below mentioned order.
For the type "trio": sample_info_file should have, affected-offspring,affected-parent,unaffected-parent trio.
For the type "duo" or "related": sample_info_file should have, affected-sample_1,affected-sample_2 duo.
For the type "recessive", a single affected sample is phased around a recessive variant using haplotype homozygosity : sample_info_file should have, affected-sample, recessive.
The function works similarly for "duo" and "related".}

\item{n.cores}{Number of cores to parallelize, default is 1 (type \code{"numeric"}).}
}
\value{
Write a valid new VCF file in output_DIR "h1" column contains the disease haplorype saved as homozygous genotypes.
}
\description{
Takes a VCF file with known disease individuals and create disease haplotypes based on trio/duo, related samples or recessive variants .
}
\details{
Make sure the input_vcf chromosome column doesn't have the "chr" prefix.
}
\examples{
orig_DIR <- getwd()
temp_DIR <- tempdir()
setwd(temp_DIR)
library(vcfR)
write.vcf(FAME1_disease_cohort,paste0(temp_DIR,"/","FAME1_disease_cohort.vcf.gz"))
sample_info=data.frame(rbind(c("HG00362_1_HG00362_1,HG00362_2_HG00362_2","duo"),c("NA11920_NA11920,Affected_parent_NA11920_Affected_parent_NA11920,Unaffected_parent_NA11920_Unaffected_parent_NA11920","trio"),c("HG00313_1_HG00313_1,HG00313_2_HG00313_2","duo")))
write.table(sample_info,paste0(temp_DIR,"/","sample_info.txt"),sep ="\t",quote=FALSE, row.names=FALSE,col.names = FALSE)
Phasing_by_pedigree(input_vcf = paste0(temp_DIR,"/FAME1_disease_cohort",".vcf.gz"),
                         output_DIR = temp_DIR,
                         sample_info_file = paste0(temp_DIR,"/","sample_info.txt"),n.cores=1)
setwd( orig_DIR )
}
